requirements.txt
numpy>=2.0.2
matplotlib>=3.10.0
pandas>=2.2.3
scikit-learn>=1.6.0
tensorflow-2.18.0
#imbalanced-learn>=0.13.0
seaborn>=0.13.2
openai>=1.58.1